Refactored configuration helpers to a new python file
