+++ .gitignore
@@ -109,3 +109,5 @@ venv.bak/
 
 # idea
 .idea/*
+
+sample-out.png
